Fix build warnings on Windows

+++ src/formatter.cc
@@ -211,7 +211,7 @@ int Formatter::parseDurationFromDecimal(const std::string value) {
     std::string numbers = value.substr(0, pos);
     double hours = 0;
     if (Poco::NumberParser::tryParseFloat(numbers, hours)) {
-      return hours * 60 * 60;
+      return int(hours * 60 * 60);
     }
   }
 
@@ -220,7 +220,7 @@ int Formatter::parseDurationFromDecimal(const std::string value) {
     std::string numbers = value.substr(0, pos);
     double hours = 0;
     if (Poco::NumberParser::tryParseFloat(numbers, hours)) {
-      return hours * 60 * 60;
+      return int(hours * 60 * 60);
     }
   }
 
@@ -230,7 +230,7 @@ int Formatter::parseDurationFromDecimal(const std::string value) {
     std::string numbers = value.substr(0, pos);
     double minutes = 0;
     if (Poco::NumberParser::tryParseFloat(numbers, minutes)) {
-      return minutes * 60;
+      return int(minutes * 60);
     }
   }
 
@@ -240,7 +240,7 @@ int Formatter::parseDurationFromDecimal(const std::string value) {
     std::string numbers = value.substr(0, pos);
     double seconds = 0;
     if (Poco::NumberParser::tryParseFloat(numbers, seconds)) {
-      return seconds;
+      return int(seconds);
     }
   }
 
@@ -254,9 +254,9 @@ int Formatter::parseDurationFromDecimal(const std::string value) {
       numbers = value.substr(pos+1, std::string::npos);
       double minutes = 0;
       if (Poco::NumberParser::tryParseFloat(numbers, minutes)) {
-        return (hours * 60 * 60) + (minutes * 60);
+        return int((hours * 60 * 60) + (minutes * 60));
       }
-      return hours * 60 * 60;
+      return int(hours * 60 * 60);
     }
   }
 
@@ -266,7 +266,7 @@ int Formatter::parseDurationFromDecimal(const std::string value) {
     std::string numbers = value.substr(0, pos);
     double minutes = 0;
     if (Poco::NumberParser::tryParseFloat(numbers, minutes)) {
-      return minutes * 60;
+      return int(minutes * 60);
     }
   }
 
@@ -276,7 +276,7 @@ int Formatter::parseDurationFromDecimal(const std::string value) {
     std::string numbers = value.substr(0, pos);
     double seconds = 0;
     if (Poco::NumberParser::tryParseFloat(numbers, seconds)) {
-      return seconds;
+      return int(seconds);
     }
   }
 
@@ -291,7 +291,7 @@ int Formatter::parseDurationFromDecimal(const std::string value) {
   // 1,5
   double hours = 0;
   if (Poco::NumberParser::tryParseFloat(value, hours)) {
-    return hours * 60 * 60;
+    return int(hours * 60 * 60);
   }
 
   return 0;
